refactor navigation logic, moved out of view logic to seperate logic
