+++ app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/ui/navigation/Routes.kt
@@ -1,5 +1,13 @@
 package io.github.viabachelora23michaelkutaibakasper.bprapp.ui.navigation
 
+import androidx.navigation.NavController
+
+
+fun navigateTo(screen: String, navController: NavController) {
+    navController.navigate(screen)
+}
+
+
 enum class BottomNavigationScreens {
     Map,
     Recommendations,
